Egregious hack to get Jasmine-Jquery fixtures working
Carolyn and I worked for a bit on Friday trying to
get the fixtures to be seen per the specification
in jasmine-jquery-rails.  We were unsuccessful.
It appears that the issue is that the fixtures
path is not made available on the web server.

This was listed as a bug here:

https://github.com/travisjeffery/jasmine-jquery-rails/issues/4

The solution, implemented here, is to define a new
Rack middleware in an initializer
(`JasmineFixtureServer`) which serves assets out
of the physical file directory
`spec/javascripts/fixtures`.

We then `mount` that engine at the URL path of
`spec/javascripts/fixtures`.  Thus we can "serve"
our fixtures (yay).

This seems like a bug to me on jquery-rails, but
this will keep things going.